Refactor methods to use helper methods for getting data for a specific week and for formatting data to assign it as a value to a date property
