+++ index.html
@@ -81,6 +81,7 @@
     <script type="text/javascript" src="./data/hydration.js"></script>
     <script type="text/javascript" src="./data/sleep.js"></script>
     <script type="text/javascript" src="./data/activity.js"></script>
+    <script type="text/javascript" src="./helpers/retrieveDataByWeek.js"></script>
     <script type="text/javascript" src="./src/UserSleep.js"></script>
     <script type="text/javascript" src="./src/AllUserSleep.js"></script>
     <script type="text/javascript" src="./src/User.js"></script>
