+++ src/AllUserSleep.js
@@ -1,8 +1,4 @@
-var dayjs = require("dayjs");
-var duration = require('dayjs/plugin/duration')
-dayjs.extend(duration);
-var isBetween = require('dayjs/plugin/isBetween')
-dayjs.extend(isBetween)
+var retrieveAllUserDataByWeek = require("./helpers/retrieveDataByWeek");
 
 class AllUserSleep {
   constructor(sleepData) {
@@ -15,19 +11,6 @@ class AllUserSleep {
     }, 0)
     const avgSleepQuality = totalSleepQuality / sleepData.length
     return avgSleepQuality
-  }
-
-  retrieveAllUserDataByWeek(date) {
-    const day1 = dayjs(new Date(date));
-    const day7 = dayjs(day1).add(dayjs.duration({"weeks" : 1}))
-    const dataForDates = this.sleepData.reduce((total, dataPoint) => {
-      if (dayjs(dataPoint.date).isBetween(day1, day7, null, "[]")) {
-        return [...total, dataPoint]
-      }
-      return total
-    }, [])
-    return dataForDates
-    //returns data containing only relevant dates in week window
   }
 
   retrieveUniqueUserIDs(weekData) {
@@ -43,7 +26,7 @@ class AllUserSleep {
   }
 
   calcAboveAvgSleepQuality(date) {
-    const dataForWeek = this.retrieveAllUserDataByWeek(date);
+    const dataForWeek = retrieveAllUserDataByWeek(this.sleepData, date);
     const userIDs = this.retrieveUniqueUserIDs(dataForWeek);
 
     //iterates through unique ids and then filters the data for the week to grab all of the data relevant to that unique id
